Move ref to TraceQLEditor
